Move edit title, delete challenege and close participation option to settings
