+++ frontend_v2/src/app/components/challenge/challengesettings/challengesettings.component.ts
@@ -5,6 +5,7 @@ import { ApiService } from '../../../services/api.service';
 import { GlobalService } from '../../../services/global.service';
 import {COMMA, ENTER} from '@angular/cdk/keycodes';
 import {MatChipInputEvent} from '@angular/material/chips';
+import { Router } from '@angular/router';
 
 @Component({
   selector: 'app-challengesettings',
@@ -17,6 +18,16 @@ export class ChallengesettingsComponent implements OnInit {
    * Challenge object
    */
   challenge: any;
+
+  /**
+   * Is challenge host
+   */
+  isChallengeHost = false;
+
+  /**
+   * To call the API inside modal for editing the challenge details
+   */
+  apiCall: any;
 
   /**
    * Participants banned emails ids
@@ -57,13 +68,16 @@ export class ChallengesettingsComponent implements OnInit {
   isBannedEmailInputVisible: boolean;
 
   constructor(private challengeService: ChallengeService, private globalService: GlobalService,
-              private apiService: ApiService, private endpointsService: EndpointsService) { }
+              private apiService: ApiService, private endpointsService: EndpointsService, private router: Router) { }
 
   ngOnInit() {
     this.challengeService.currentChallenge.subscribe(
       challenge => {
         this.challenge = challenge;
         this.updateView();
+    });
+    this.challengeService.isChallengeHost.subscribe(status => {
+      this.isChallengeHost = status;
     });
   }
 
@@ -155,4 +169,135 @@ export class ChallengesettingsComponent implements OnInit {
         () => {}
       );
   }
+
+  /**
+   * Close participation function
+   */
+  stopParticipation(event) {
+    event.preventDefault();
+    const participationState = (this.challenge['is_registration_open']) ? 'Close' : 'Open';
+
+    this.apiCall = () => {
+      if (this.isChallengeHost && this.challenge['id'] !== null) {
+        const BODY = JSON.stringify({
+          'is_registration_open': !this.challenge['is_registration_open']
+        });
+        this.apiService.patchUrl(
+          this.endpointsService.editChallengeDetailsURL(this.challenge.creator.id, this.challenge.id),
+          BODY
+        ).subscribe(
+          () => {
+            this.challenge['is_registration_open'] = !this.challenge['is_registration_open'];
+            this.globalService.showToast(
+              'success', 'Participation is ' + participationState.replace('n', 'ne') + 'd successfully', 5
+            );
+          },
+          err => {
+            this.globalService.handleApiError(err, true);
+            this.globalService.showToast('error', err);
+          },
+          () => {}
+          );
+      }
+    };
+
+    const PARAMS = {
+      title: participationState + ' participation in the challenge?',
+      content: '',
+      confirm: 'Yes, I\'m sure',
+      deny: 'No',
+      confirmCallback: this.apiCall
+    };
+    this.globalService.showConfirm(PARAMS);
+  }
+
+  /**
+   * Edit challenge title function
+   */
+  editChallengeTitle() {
+    const SELF = this;
+
+    SELF.apiCall = (params) => {
+      const BODY = JSON.stringify(params);
+      SELF.apiService.patchUrl(
+        SELF.endpointsService.editChallengeDetailsURL(SELF.challenge.creator.id, SELF.challenge.id),
+        BODY
+        ).subscribe(
+        data => {
+          SELF.challenge.title = data.title;
+          SELF.globalService.showToast('success', 'The challenge title is  successfully updated!', 5);
+        },
+        err => {
+          SELF.globalService.handleApiError(err, true);
+          SELF.globalService.showToast('error', err);
+        },
+        () => console.log('EDIT-CHALLENGE-TITLE-FINISHED')
+      );
+    };
+
+    const PARAMS = {
+      title: 'Edit Challenge Title',
+      content: '',
+      confirm: 'Submit',
+      deny: 'Cancel',
+      form: [
+        {
+          name: 'editChallengeTitle',
+          isRequired: true,
+          label: 'title',
+          placeholder: 'Challenge Title',
+          type: 'text',
+          value: this.challenge.title
+        },
+      ],
+      confirmCallback: SELF.apiCall
+    };
+    SELF.globalService.showModal(PARAMS);
+  }
+
+  /**
+   * Delete challenge
+   */
+  deleteChallenge() {
+    const SELF = this;
+    const redirectTo = '/dashboard';
+
+    SELF.apiCall = () => {
+      const BODY = JSON.stringify({});
+      SELF.apiService.postUrl(
+        SELF.endpointsService.deleteChallengeURL(SELF.challenge.id),
+        BODY
+        ).subscribe(
+        data => {
+          SELF.router.navigate([redirectTo]);
+          SELF.globalService.showToast('success', 'The Challenge is successfully deleted!', 5);
+        },
+        err => {
+          SELF.globalService.handleApiError(err, true);
+          SELF.globalService.showToast('error', err);
+        },
+        () => console.log('DELETE-CHALLENGE-FINISHED')
+      );
+    };
+
+    const PARAMS = {
+      title: 'Delete Challenge',
+      content: '',
+      confirm: 'I understand consequences, delete the challenge',
+      deny: 'Cancel',
+      form: [
+        {
+          name: 'challegenDeleteInput',
+          isRequired: true,
+          label: '',
+          placeholder: 'Please type in the name of the challenge to confirm',
+          type: 'text',
+          value: ''
+        },
+      ],
+      confirmCallback: SELF.apiCall
+    };
+    SELF.globalService.showModal(PARAMS);
+  }
+
 }
